moved some files around, code cleanup, making sure things look neat
